ID: uint -> uint64

+++ note.go
@@ -1,7 +1,7 @@
 package main
 
 type Note struct {
-	ID uint `gorm:"primaryKey" json:"id"`
+	ID uint64 `gorm:"primaryKey" json:"id"`
 	Title string `json:"title"`
 	Content string `json:"content"`
 }
